第８章まで

+++ ZKNOTE/app/Http/Controllers/ArticleController.php
@@ -44,6 +44,7 @@ class ArticleController extends Controller
             'body' => 'required'
         ]);
         $article = new Article();
+        $article->user_id = \Auth::id(); // 一行追加
         $article->title = $request->title;
         $article->body = $request->body;
         $article->save();
@@ -71,6 +72,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        $this->authorize($article);
         $data = ['article' => $article];
         return view('articles.edit', $data);
     }
@@ -84,6 +86,7 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        $this->authorize($article);
         $this->validate($request, [
             'title' => 'required|max:255',
             'body' => 'required'
@@ -102,6 +105,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        $this->authorize($article);
         $article->delete();
         return redirect(route('articles.index'));
     }
